refactor(storage): extract Storage interface to storagetypes package to break import cycle
